+++ CHANGELOG.md
@@ -34,6 +34,8 @@
   ([#6087](https://github.com/mitmproxy/mitmproxy/pull/6087), @sujaldev)
 * focus-follow shortcut will now work in flow view context too.
   ([#6088](https://github.com/mitmproxy/mitmproxy/pull/6088), @sujaldev)
+* Fix a bug where a server connection timeout would cause requests to be issued with a wrong SNI in reverse proxy mode.
+  ([#6148](https://github.com/mitmproxy/mitmproxy/pull/6148), @mhils)
 
 ### Breaking Changes